compat: cast enums to int

+++ avs_core/core/internal.h
@@ -47,10 +47,10 @@
 
 enum MANAGE_CACHE_KEYS
 {
-  MC_RegisterCache     = 0xFFFF0004,
-  MC_UnRegisterCache   = 0xFFFF0006,
-  MC_NodCache          = 0xFFFF0007,
-  MC_NodAndExpandCache = 0xFFFF0008,
+  MC_RegisterCache     = (int)0xFFFF0004,
+  MC_UnRegisterCache   = (int)0xFFFF0006,
+  MC_NodCache          = (int)0xFFFF0007,
+  MC_NodAndExpandCache = (int)0xFFFF0008,
   MC_RegisterMTGuard,
   MC_UnRegisterMTGuard
 };
